smol refactor
just moved the rei stuff into a rei folder instead of leaving it in the compat folder
also removed the test transmute recipe
