+++ src/main/java/com/github/ethanicuss/astraladditions/compat/rei/desizer/DesizerCategory.java
@@ -1,6 +1,6 @@
-package com.github.ethanicuss.astraladditions.compat.desizer;
+package com.github.ethanicuss.astraladditions.compat.rei.desizer;
 
-import com.github.ethanicuss.astraladditions.compat.AstralAdditionsREIClientPlugin;
+import com.github.ethanicuss.astraladditions.compat.rei.AstralAdditionsREIClientPlugin;
 import com.github.ethanicuss.astraladditions.registry.ModBlocks;
 import com.google.common.collect.Lists;
 import me.shedaniel.math.Point;
