+++ src/main/java/com/github/ethanicuss/astraladditions/compat/rei/transmute/TransmuteCategory.java
@@ -1,7 +1,7 @@
-package com.github.ethanicuss.astraladditions.compat.transmute;
+package com.github.ethanicuss.astraladditions.compat.rei.transmute;
 
-import com.github.ethanicuss.astraladditions.compat.AstralAdditionsREIClientPlugin;
-import com.github.ethanicuss.astraladditions.compat.widgets.RotatedArrowWidget;
+import com.github.ethanicuss.astraladditions.compat.rei.AstralAdditionsREIClientPlugin;
+import com.github.ethanicuss.astraladditions.compat.rei.widgets.RotatedArrowWidget;
 import com.github.ethanicuss.astraladditions.fluids.ModFluids;
 import com.google.common.collect.Lists;
 import me.shedaniel.math.Point;
