+++ src/main/java/com/github/ethanicuss/astraladditions/compat/rei/vacuum/VacuumCategory.java
@@ -1,7 +1,7 @@
-package com.github.ethanicuss.astraladditions.compat.vacuum;
+package com.github.ethanicuss.astraladditions.compat.rei.vacuum;
 
-import com.github.ethanicuss.astraladditions.compat.AstralAdditionsREIClientPlugin;
-import com.github.ethanicuss.astraladditions.compat.widgets.RotatedArrowWidget;
+import com.github.ethanicuss.astraladditions.compat.rei.AstralAdditionsREIClientPlugin;
+import com.github.ethanicuss.astraladditions.compat.rei.widgets.RotatedArrowWidget;
 import com.github.ethanicuss.astraladditions.registry.ModItems;
 import com.google.common.collect.Lists;
 import me.shedaniel.math.Point;
