--- src/main/resources/data/astraladditions/recipes/transmute.json
@@ -1,21 +1,0 @@
-{
-  "type": "astraladditions:shimmer_transmute",
-  "input": {
-    "item": "minecraft:iron_ingot",
-    "count": 2
-  },
-  "output": [
-    {
-      "item": "minecraft:diamond",
-      "count": 1
-    },
-    {
-      "item": "minecraft:iron_ingot",
-      "count": 2
-    },
-    {
-      "item": "minecraft:dirt",
-      "count": 1
-    }
-  ]
-}
